Fix crash on calling callback of navigator.mozApps.install
We were reusing the JSContextRef in the call to mozapps_install,
to call the callbacks. This is wrong, as this context may not exist
anymore (in fact it does not exist).

Now we set the global context in private area of mozApps object,
so that we can use it for the call.